Added countdown background & snow effect
